+++ TombEngine/Scripting/Internal/ReservedScriptNames.h
@@ -348,6 +348,8 @@ static constexpr char ScriptReserved_KeyClearAll[]					= "KeyClearAll";
 static constexpr char ScriptReserved_FlipMap[]						= "FlipMap";
 static constexpr char ScriptReserved_GetFlipMapStatus[]				= "GetFlipMapStatus";
 static constexpr char ScriptReserved_PlayFlyBy[]					= "PlayFlyBy";
+static constexpr char ScriptReserved_GetFlyByPosition[]				= "GetFlyByPosition";
+static constexpr char ScriptReserved_GetFlyByRotation[]				= "GetFlyByRotation";
 
 static constexpr char ScriptReserved_PlayCamera[]					= "PlayCamera";
 static constexpr char ScriptReserved_ResetObjCamera[]				= "ResetObjCamera";