+++ CHANGELOG.md
@@ -29,10 +29,11 @@ TombEngine releases are located in this repository (alongside with Tomb Editor):
   - You must use this version: https://github.com/TombEngine/Resources/raw/refs/heads/main/Wad2%20Objects/Interactables/TEN_Waterfall_Emitter.wad2
 
 ### Lua API changes
-* Added Lerp() function to the Rotation object to allow linear interpolation between rotations.
 * Added diary module.
+* Added View.GetFlyByPosition() and View.GetFlyByRotation() functions to get flyby sequence parameters at a specified time point.
 * Added Effects.EmitAirBubble() function to spawn air bubbles.
 * Added additional arguments for Sprite object slot and starting rotation value for EmitParticle function.
+* Added Lerp() function to the Rotation object to allow linear interpolation between rotations.
 * Added various Translate() methods to Vec2 and Vec3 script objects.
   
 ## [Version 1.7.1](https://github.com/TombEngine/TombEditorReleases/releases/tag/v1.7.4) - 2025-04-01